// packages/shared/structured-data/src/types.ts
export type JsonLd = Record<string, unknown>;
export type JsonLdNodeType =
  | "WebPage"
  | "WebSite"
  | "Organization"
  | "SoftwareApplication"
  | "WebApplication"
  | "Article"
  | "TechArticle"
  | "BlogPosting"
  | "BreadcrumbList"
  | "FAQPage"
  | "QAPage"
  | "HowTo"
  | "ItemList"
  | "SoftwareSourceCode"
  | "Product"
  | "Dataset"
  | "Event"
  | "VideoObject"
  | "ImageObject"
  | "ProfilePage"
  | "Review"
  | "AggregateRating"
  | "Course"
  | "CourseInstance"
  | "DiscussionForumPosting"
  | "Book"
  | "ReadAction"
  | "ClaimReview"
  | "EmployerAggregateRating"
  | "MonetaryAmountDistribution"
  | "JobPosting"
  | "LocalBusiness"
  | "MemberProgram"
  | "MathSolver"
  | "MerchantReturnPolicy"
  | "OfferShippingDetails"
  | "Movie"
  | "ProductGroup"
  | "Recipe"
  | "SpeakableSpecification"
  | "VacationRental"
  | "Vehicle";

export interface StructuredDataImage {
  src?: string;
  url?: string;
  width?: number;
  height?: number;
  caption?: string;
  alt?: string;
}